requirements.txt
langchain
langchain-community
openai
faiss-cpu
pdfplumber
pytesseract
Pillow
python-dotenv
tiktoken
streamlit